Extract classes and enums definitions into their own files
