+++ cip_enip_driver/CommandSpecificDataRegisterSession.cs
@@ -1,0 +1,16 @@
+// -----------------------------------------------------------------------------------------------------------------
+// Project: CIP driver over Ethernet/IP - Copyright (C) 2019, Bruno Ataíde (Techsteel), All rights reserved
+// License: GNU General Public License 3.0
+// Public Repo: https://github.com/Bataide/cip_enip_driver
+// Description: Structs
+// -----------------------------------------------------------------------------------------------------------------
+using System;
+
+namespace Techsteel.Drivers.CIP
+{
+    public class CommandSpecificDataRegisterSession : CIPSerializer
+    {
+        public ushort ProtocolVersion;
+        public ushort OptionsFlags;
+    }
+}
