+++ cip_enip_driver/EncapsulationCommands.cs
@@ -1,0 +1,24 @@
+// -----------------------------------------------------------------------------------------------------------------
+// Project: CIP driver over Ethernet/IP - Copyright (C) 2019, Bruno Ataíde (Techsteel), All rights reserved
+// License: GNU General Public License 3.0
+// Public Repo: https://github.com/Bataide/cip_enip_driver
+// Description: Enums
+// -----------------------------------------------------------------------------------------------------------------
+using System;
+
+namespace Techsteel.Drivers.CIP
+{
+    public enum EncapsulationCommands : ushort
+    {
+        NOP = 0x0000,
+        ListServices = 0x0004,
+        ListIdentity = 0x0063,
+        ListInterfaces = 0x0064,
+        RegisterSession = 0x0065,
+        UnRegisterSession = 0x0066,
+        SendRRData = 0x006F,
+        SendUnitData = 0x0070,
+        IndicateStatus = 0x0072,
+        Cancel = 0x0073,
+    }
+}
