+++ cip_enip_driver/SegmentDataSubType.cs
@@ -1,0 +1,16 @@
+// -----------------------------------------------------------------------------------------------------------------
+// Project: CIP driver over Ethernet/IP - Copyright (C) 2019, Bruno Ataíde (Techsteel), All rights reserved
+// License: GNU General Public License 3.0
+// Public Repo: https://github.com/Bataide/cip_enip_driver
+// Description: Enums
+// -----------------------------------------------------------------------------------------------------------------
+using System;
+
+namespace Techsteel.Drivers.CIP
+{
+    public enum SegmentDataSubType
+    {
+        SimpleDataSegment = 0,
+        ANSIExtendedSymbolSegment = 17,
+    }
+}
